requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
comm==0.2.3
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.8.15
decorator==5.2.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.63
langchain-experimental==0.3.3
langchain-neo4j==0.1.1
langchain-openai==0.2.10
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.1.147
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.48.0
neo4j==5.27.0
neo4j-graphrag==1.9.0
nest-asyncio==1.6.0
numpy==2.0.2
openai==1.56.0
orjson==3.11.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.9.1
pydantic-settings==2.10.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.19.2
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
streamlit==1.35.0
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
